refactor: update the session verification return

+++ src/features/pots/data-access.ts
@@ -25,13 +25,13 @@ export async function createPot({
   target,
   colorId,
 }: PotCreate): Promise<DALReturn<PotCreateErrors>> {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   try {
     await prisma.pot.create({
       data: {
-        userId,
+        userId: session.userId,
         name: name,
         target: target,
         colorId: colorId,
@@ -65,12 +65,12 @@ export async function updatePot({
   target,
   colorId,
 }: PotUpdate): Promise<DALReturn<PotCreateErrors>> {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   try {
     await prisma.pot.update({
-      where: { userId, id },
+      where: { userId: session.userId, id },
       data: {
         name: name,
         target: target,
@@ -100,11 +100,11 @@ export async function updatePot({
 // ============================================
 
 export async function deletePot(potId: string): Promise<{ success: boolean }> {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   try {
-    await prisma.pot.delete({ where: { id: potId, userId } })
+    await prisma.pot.delete({ where: { id: potId, userId: session.userId } })
     return { success: true }
   } catch {
     return { success: false }
@@ -119,14 +119,14 @@ export async function updatePotAmount(
   { id, amountToUpdate }: PotAmountUpdate,
   operation: "increment" | "decrement"
 ): Promise<DALReturn<PotAmountUpdateErrors>> {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   try {
     const amountToUpdateAsDecimal = new Prisma.Decimal(amountToUpdate)
 
     await prisma.pot.update({
-      where: { userId, id },
+      where: { userId: session.userId, id },
       data: { currentAmount: { [operation]: amountToUpdateAsDecimal } },
     })
     return { success: true }
@@ -166,21 +166,21 @@ export type PotsSummary = {
 }
 
 export async function getPots(take?: number) {
-  const userId = await verifySession()
-  if (!userId) redirect("/login")
+  const session = await verifySession()
+  if (!session) redirect("/login")
 
   // await new Promise((resolve) => setTimeout(resolve, 2000))
 
   const [pots, aggregates] = await prisma.$transaction([
     prisma.pot.findMany({
-      where: { userId },
+      where: { userId: session.userId },
       orderBy: { createdAt: "desc" },
       take: take,
       select: potSelect,
     }),
 
     prisma.pot.aggregate({
-      where: { userId },
+      where: { userId: session.userId },
       _sum: { currentAmount: true, target: true },
       _count: true,
     }),
